[WIP] Considerably rework `textLine2CipherStatisticsDataset.py`
The main point of these changes was the addition of rotor ciphers to
the training process. The training pipeline should now support the
calculation of new features for rotor ciphers.
One of the challenges was the combination of the existing pipeline
(starting from plaintext files) and the new pipeline for rotor ciphers
(starting with already encrypted samples).
The existing pipeline takes paths to plaintext files, uses
tensorflows TextLineDataset to load these files and iterate their
lines, encrypts the lines with all configured ciphers (with all their
different key lengths) and then calculates their features.

The new pipeline in contrast simply loads the ciphertext files at once
(can easily be changed to a lazy fashion, should the input grow from
the current 5 files of 1000 lines each) and calculates the features
for each line.

The previous implementation had some issues, which meant that it was
not sensible to just extend the existing pipeline.
These issues include:
- Dysfunctional parallel processing of inputs (encryption and feature
calculation)
- Leakage of implementation details (`TextLine2CipherStatisticsDataset`
- iterator that implements most of the existing pipeline - starts
processes, that are returned to the calling code, which must join the
processes.)
- Confusing naming and structure

Changes include:
- Add a new dataset type `CombinedCipherStatisticsDataset` (with an
iterator interface) which is responsible for concurrently getting a new
batch from the inputs (plaintexts for ACA ciphers and ciphertexts for
rotor ciphers) and performing the processing of the inputs (encryption,
feature calculation, etc.)
  - uses a processing pool
  - also just passes simple (pickable) values between processes
    - prevents the issues of the old implementation
- replace the existing `TextLine2CipherStatisticsDataset` by a
simpler iterator
  - called `PlaintextPathsDataset`
  - simply returns batched inputs
  - does not process these inputs
- also addd corresponding `RotorCiphertextDataset` for rotor ciphers
- add worker processes that only process the prevously generated
batched inputs
  - called `PlaintextLine2CipherStatisticsWorker` and
  `CiphertextLine2CipherStatisticsWorker`
- extract the long parameter list of `TextLine2CipherStatisticsDataset`
to it's own class
  - called `PlaintextDatasetParameters`
- do the same for the new `RotorCiphertextDataset`
- add a class `ConfigParams` as a data store for some of the values
of `config.py`
  - previously the global state could simply be accessed in different
processes
- add `TrainingBatch` as a simple means of improving the structure of
the code
  - previously returned batched training inputs (features and their
  labels were structured as a list of tuples of lists of same length.
  - was somewhat confusing, now has a clear name