[WIP] Correctly load plaintext datasets
The old implementation did nor work as expected. It did not combine
the datasets to one large dataset, but only used the inputs of the
dataset corresponding to the last plaintext file. Therefore only a tiny
subset of the available plaintext lines is used in the training process.

Simply used the initializer of tf.data.TextLineDataset with all paths
as input.