[WIP] Add meta-clf. pipeline and change existing pl. in train.py
Currently there are different approaches to support rotor ciphertexts
in NCID:
One approach is to simply train models with the
features of the paper by Dalton and Stamp and to take the existing
models of the NCID project and train a meta classifier, which decides
which model to use for evaluation of the given ciphertext.
Another approach is the extension of the existing models by with the
new features.

This commit adds:
- The meta-classifier pipeline, which can be used to train a meta
classifier, that decides which type of architecture should be used to
evaluate the user input.
- Support for the latest changes to
`textLine2CipherStatisticsDataset.py`. That is the training with rotor
ciphertexts as well as better multiprocessing support and other bug
fixes.